refactor: 마이페이지 자기소개 파트, 이미지, 닉네임, 소개 기능에 따른 컴포넌트 분리 #116
